requirements.txt
streamlit==1.39.0
google-genai==0.3.0
requests==2.32.3
